feat: enforce noOfPages validation and document exception handling
